engine: consolidate simlin-core and simlin-compat into simlin-engine
This consolidates the three Rust packages (simlin-core, simlin-compat, and
simlin-engine) into a single simlin-engine package. This simplifies the
codebase and prepares for future work on writing MDL files.

Key changes:
- Moved common.rs, datamodel.rs, and results.rs from simlin-core into simlin-engine
- Moved mdl/, xmile.rs and compat functions from simlin-compat into simlin-engine
- Moved all integration tests from simlin-compat/tests/ to simlin-engine/tests/
- Added new features to simlin-engine: file_io (for CSV support) and xmutil
- xmutil remains an optional dependency, only enabled for tests/benchmarks
- Updated libsimlin to depend only on simlin-engine
- Updated simlin-cli to use simlin-engine with file_io feature
- Removed simlin-core and simlin-compat from workspace and deleted their directories

The xmutil C++ dependency is kept as an optional feature that is NEVER
enabled for libsimlin builds, ensuring that no C++ code ends up in the
WASM builds or FFI library.
