+++ src/libsimlin/Cargo.toml
@@ -15,7 +15,6 @@ default = []
 
 [dependencies]
 simlin-engine = { version = "0.1", path = "../simlin-engine", default-features = false }
-simlin-compat = { version = "0.1", path = "../simlin-compat", default-features = false }
 prost = "0.14"
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
