+++ src/simlin-cli/Cargo.toml
@@ -12,5 +12,5 @@ path = "src/main.rs"
 [dependencies]
 pico-args = "0.5"
 stringreader = "0.1"
-simlin-compat = { version = "0.1", path = "../simlin-compat", features = ["file_io"] }
+simlin-engine = { version = "0.1", path = "../simlin-engine", features = ["file_io"] }
 simlin = { version = "0.1", path = "../libsimlin" }
