--- src/simlin-compat/Cargo.toml
@@ -1,45 +1,0 @@
-[package]
-name = "simlin-compat"
-version = "0.1.0"
-authors = ["Bobby Powers <bobbypowers@gmail.com>"]
-license = "MIT"
-edition = "2024"
-
-[lib]
-crate-type = ["rlib"]
-
-[features]
-default = ["file_io"]
-xmutil = ["dep:xmutil"]
-file_io = ["dep:csv"]
-
-[dependencies]
-csv = { version = "1", optional = true }
-float-cmp = "0.10"
-quick-xml = { version = "0.38", features = [ "serialize", "overlapped-lists" ] }
-serde = { version = "1", features = [ "derive" ] }
-simlin-core = { version = "0.1", path = "../simlin-core" }
-simlin-engine = { version = "0.1", path = "../simlin-engine", default-features = false }
-xmutil = { version = "1", path = "../xmutil", optional = true }
-
-[dev-dependencies]
-simlin-engine = { version = "0.1", path = "../simlin-engine", features = ["ai_info"] }
-test-generator = { version = "0.3.0", git = "https://github.com/bpowers/test-generator", rev = "b78145bfb6a6f81425dfd6fbacb9c03624e79b2c" }
-ed25519-dalek = "2"
-ssh-key = "0.6"
-base64 = "0.22"
-serde_json = "1"
-criterion = { version = "0.5", features = ["html_reports"] }
-dhat = "0.3.3"
-
-[[example]]
-name = "profile_clearn"
-required-features = ["xmutil"]
-
-[[example]]
-name = "dhat_profile"
-
-[[bench]]
-name = "clearn_parse"
-harness = false
-required-features = ["xmutil"]
