+++ src/xmutil/Cargo.toml
@@ -17,4 +17,3 @@ lazy_static = "1"
 cc = "1"
 
 [dev-dependencies]
-simlin-compat = { version = "0.1", path = "../simlin-compat" }
